composite and composer together in composer.py
